Feat: 상품 리뷰 조회시 옵션 이름 추가

+++ src/main/java/site/keydeuk/store/entity/OrderItem.java
@@ -14,7 +14,7 @@ public class OrderItem {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
     private Integer count;
-    private Integer switchOptionId;
+    private Long switchOptionId;
     @ManyToOne
     @JoinColumn(name = "order_id")
     private Order order;
